combine decorators
